convert get-stylesheet to ts
